feat: move PinListState to commons, add bookmark/pin state to IAccount
